+++ src/tools/index.ts
@@ -2,7 +2,6 @@
  *  Tool Module Exports
  *--------------------------------------------------------------------------------------------*/
 
-export { DelegateToAgentTool } from "./delegateToAgent";
 export { GeminiSearchTool } from "./geminiSearch";
 export { MiniMaxSearchTool } from "./minimaxSearch";
 export { registerAllTools } from "./registry";